+++ CHANGELOG.rst
@@ -11,7 +11,7 @@ Added
 
 * Allow equal sign as separator in set directive for better readability.
 
-* Allow macro name in enddef directives.
+* Allow macro names in enddef and endcall directives.
 
 * Implement assert directive.